+++ client/src/components/habits/Habits/HabitEntrySlider.tsx
