+++ client/src/components/habits/Habits/style/Completion.style.ts
